Rename color blind themes (#254)
* Rename protanopia themes to colorblind

* Rename protanopia themes to colorblind

* Create green-apples-hang.md

* Update green-apples-hang.md
