+++ .changeset/green-apples-hang.md
@@ -1,0 +1,5 @@
+---
+"@primer/primitives": major
+---
+
+Rename color blind themes
