+++ data/colors/index.ts
@@ -1,15 +1,15 @@
 import light from './themes/light'
-import light_protanopia from './themes/light_protanopia'
+import light_colorblind from './themes/light_colorblind'
 import dark from './themes/dark'
 import dark_dimmed from './themes/dark_dimmed'
 import dark_high_contrast from './themes/dark_high_contrast'
-import dark_protanopia from './themes/dark_protanopia'
+import dark_colorblind from './themes/dark_colorblind'
 
 export default {
   light,
-  light_protanopia,
+  light_colorblind,
   dark,
   dark_dimmed,
   dark_high_contrast,
-  dark_protanopia
+  dark_colorblind
 }
